requirements.txt
Django
torch
torchvision
Pillow